+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.7.2] - 23.04.2023
+
+### Added
+
+- Module `multiprocessing`
+- Method `dict_utils.dict_intersection`
+
 ## [0.7.1] - 23.04.2023
 
 ### Added